Extreact DOM template from jquery.template
