+++ js/ui/set_template_engine.js
@@ -22,4 +22,4 @@
 * @export default
 */
 
-module.exports = require("./widget/jquery.template").setTemplateEngine;
+module.exports = require("./widget/template_engine").setTemplateEngine;
