+++ testing/tests/DevExpress.ui/widget.tests.js
@@ -3,7 +3,7 @@ var $ = require("jquery"),
     registerComponent = require("core/component_registrator"),
     devices = require("core/devices"),
     TemplateBase = require("ui/widget/ui.template_base"),
-    Template = require("ui/widget/jquery.template"),
+    Template = require("ui/widget/template"),
     DataHelperMixin = require("data_helper"),
     DataSource = require("data/data_source/data_source").DataSource,
     keyboardMock = require("../../helpers/keyboardMock.js"),
